split exploration tabs into separate test files
